MAINT: Handle gaps in input to daily bars writer
Previously, a dataframe pass into BcolzDailyBarWriter.write that was
missing an expected session between its first and last sessions would be
written incorrectly. Upon converting the dataframe to a ctable, the
values for all days following the gap would be shifted backwards, and
nans would be shifted in at the end.

This commit handles the issue by reindexing the input dataframe against
the expected sessions, which will fill any gaps with nans at the
expected index. Also adds BcolzDailyBarWriterMissingDataTestCase to
catch any regressions.